xdsl: Move BinaryOperation, more refactoring
