+++ xdsl/dialects/arith.py
@@ -5,7 +5,7 @@ from typing import Union
 
 from xdsl.dialects.builtin import (ContainerOf, Float16Type, Float64Type, IndexType,
                                    IntegerType, Float32Type, IntegerAttr)
-from xdsl.ir import MLContext, Operation, BinaryOperation, SSAValue
+from xdsl.ir import MLContext, Operation, SSAValue
 from xdsl.irdl import (AnyOf, irdl_op_definition, AttributeDef, AnyAttr,
                        ResultDef, OperandDef, Attribute)
 from xdsl.utils.exceptions import VerifyException
@@ -80,6 +80,20 @@ class Constant(Operation):
         return Constant.create(
             result_types=[typ],
             attributes={"value": IntegerAttr.from_params(val, typ)})
+
+
+@dataclass
+class BinaryOperation(Operation):
+    """A generic operation. Operation definitions inherit this class."""
+
+    # TODO replace with trait
+    def verify_(self) -> None:
+        if self.lhs.typ != self.rhs.typ or self.rhs.typ != self.result.typ:
+            raise VerifyException(
+                "expect all input and result types to be equal")
+
+    def __hash__(self) -> int:
+        return id(self)
 
 
 @irdl_op_definition
@@ -455,9 +469,6 @@ class Cmpi(Operation):
             operands=[operand1, operand2],
             result_types=[IntegerType.from_width(1)],
             attributes={"predicate": IntegerAttr.from_int_and_width(arg, 64)})
-
-    def verify_(self) -> None:
-        pass
 
     @staticmethod
     def from_mnemonic(operand1: Union[Operation, SSAValue],
